fix 311 line table entries and remove unused code
