+++ xdis/opcodes/opcode_313.py
@@ -341,7 +341,7 @@ for fmt_table in (opcode_arg_fmt, opcode_extended_fmt):
 opcode_arg_fmt13 = opcode_arg_fmt
 opcode_extended_fmt13 = opcode_extended_fmt
 
-from xdis.opcodes.opcode_312 import findlinestarts as findlinestarts_312, parse_location_entries, format_CALL_INTRINSIC_1, format_CALL_INTRINSIC_2
+from xdis.opcodes.opcode_312 import findlinestarts as findlinestarts_312
 
 # update any calls to findlinestarts to include the version tuple
 def findlinestarts_313(code, dup_lines=False):
